File structure cleanup
All the giin_* functions are in a lib sub-directory. The inpainting
processing is now done by the inpaint function. This function saves
results which can then be visualized by the visualize function. There
are no more images in the repository, they have to be put in the data
sub-directory. A wrapper shell script launch.sh can be used to launch
any function from the command line. This is useful for processing on the
cluster.